Dockerize and add Google Analytics
Episode 3 part 1. We created the dockerfile and ignore file for deploying the website as a docker image. This is the pre-requisite for hosting on GCloud cloud run.

Add google analytics tags for the website to track website behavior.

Update readme with docker build and run commands.

Testing:
Built the image and ran it locally. Tested on localhost homepage.
Verified that google analytics metrics were being emitted and picked up in google analytics console.